// src/enums/abilityId.ts
export enum AbilityId {
  TRIPLE_REND = "TRIPLE_REND",
  PIERCING_THRUST = "PIERCING_THRUST",
  REAPERS_EDGE = "REAPERS_EDGE",
  MEND_WOUNDS = "MEND_WOUNDS",
  POWER_SURGE = "POWER_SURGE",
  GUARDIANS_BOON = "GUARDIANS_BOON",
  DRAIN_ENERGY = "DRAIN_ENERGY",
  WEAKEN_RESOLVE = "WEAKEN_RESOLVE",
  CRUSH_ARMOR = "CRUSH_ARMOR",
  PILLAGE = "PILLAGE",
  NULLIFY = "NULLIFY",
  DISEMPOWER = "DISEMPOWER",
}